adjust server ui
